backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
motor==3.5.1
pymongo==4.8.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Validation and settings
pydantic==2.9.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.2.0

# CORS
python-multipart==0.0.9
